Rename W3CActions to just Actions
